Test(refactor): should_return_the_sum_of_numbers_contains_in_string_separated_by_commas

+++ src/core/stringCalculator.ts
@@ -1,6 +1,6 @@
 export const add = (theOperationInThis: string) => {
     const theOperationInThisToIterate = theOperationInThis.split((","))
-    const sum = theOperationInThisToIterate.reduce((previousNumber, currentNumber) => previousNumber + parseInt(currentNumber), 0)
+    const sum = theOperationInThisToIterate.reduce((previousNumber, currentNumber) => previousNumber + parseFloat(currentNumber), 0)
 
     return theOperationInThis === "" ? "0" : sum.toString()
 }
